refactor(server): separate user's favourites
UsersModule previously contained all of the user's favourites
controllers and handlers making it very cluttered and hard
to maintain. This commit moves all of the logic into a newly
created FavouritesModule, making code easier to maintain and
providing additional encapsulation thanks to usage of facade pattern.
