+++ packages/server/src/favourites/api/index.ts
@@ -1,0 +1,2 @@
+export * from './FavouriteDishesController';
+export * from './FavouriteRestaurantsController';
