+++ packages/server/src/favourites/domain/index.ts
