+++ packages/server/src/favourites/index.ts
@@ -1,0 +1,1 @@
+export * from './FavouritesModule';
