--- packages/server/src/users/favourites/api/index.ts
@@ -1,2 +1,0 @@
-export { FavouriteDishesController } from './FavouriteDishesController';
-export { FavouriteRestaurantsController } from './FavouriteRestaurantsController';
