+++ packages/server/src/users/infra/index.ts
@@ -1,1 +1,2 @@
+export * from './UserFavFacade';
 export * from './UsersFacade';
